[Feat] : Repo 모델 추가
Git: #1
