+++ GitHubSearchApp_iOS_TCA/App/RootView.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import ComposableArchitecture
 
 struct RootView: View {
   var body: some View {
