+++ Tests/GitHubSearchAppComponentTests/GitHubSearch/Store/GitHubSearchStoreTests.swift
@@ -1,0 +1,14 @@
+//
+//  GitHubSearchStoreTests.swift
+//  GitHubSearchAppComponentTests
+//
+//  Created by minii on 2023/02/24.
+//
+
+import XCTest
+import SwiftUI
+import ComposableArchitecture
+
+final class GitHubSearchStoreTests: XCTestCase {
+  
+}
